% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{get_album_tracks}
\alias{get_album_tracks}
\title{Get all tracks from a given album}
\usage{
get_album_tracks(album_id, album_mbid = NULL, has_lyrics = TRUE, page = 1,
  page_size = 25)
}
\arguments{
\item{album_id}{numeric. Musixmatch album id.}

\item{album_mbid}{character. Musicbrainz album id. Defaults to NULL.}

\item{has_lyrics}{logical. Only return tracks with lyrics? Defaults to TRUE.}

\item{page}{numeric. The page number to return for paginated results. Defaults
to 1.}

\item{page_size}{numeric. The number of results to return per page. Defaults
to 25.}
}
\value{
A tibble containing tracks from the given album.
}
\description{
\code{get_album_tracks} returns a tibble of all tracks and associated details
for a given album.
}
\examples{
\dontrun{
# Get tracks from Taylor Swift's reputation album
get_album_tracks(26806100)
}

}
\references{
\url{https://developer.musixmatch.com/documentation/api-reference/album-tracks-get}
}
