% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album.R
\name{get_artist_albums}
\alias{get_artist_albums}
\title{Get albums for a specific artist}
\usage{
get_artist_albums(artist_id, artist_mbid = NULL, group_by = 1,
  sort = c("desc", "asc"), page = 1, page_size = 50)
}
\arguments{
\item{artist_id}{numeric. Musixmatch artist id.}

\item{artist_mbid}{character. Musicbrainz artist id. Defaults to NULL.}

\item{group_by}{numeric. Group by album name.}

\item{sort}{character. Sort by album release date. Either ascending (asc)
or descending (desc).}

\item{page}{numeric. Page of results to return. Defaults to 1.}

\item{page_size}{numeric. Number of results per page. Defaults to 50.}
}
\value{
A tibble of album details for the provided artist.
}
\description{
\code{get_artist_albums} returns all the albums for a given artist and details
about each album.
}
\examples{
\dontrun{
# Get all of Taylor Swift's albums
get_artist_albums(259675)
}

}
\references{
\url{https://developer.musixmatch.com/documentation/api-reference/artist-albums-get}
}
