% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{get_track_lyrics}
\alias{get_track_lyrics}
\title{Get lyrics for a given track}
\usage{
get_track_lyrics(track_id, commontrack_id = NULL)
}
\arguments{
\item{track_id}{numeric. Musixmatch track id.}

\item{commontrack_id}{numeric. Musixmatch commontrack id. Defaults to NULL.}
}
\value{
A tibble containing lyrics and lyric metadata for the requested track.
}
\description{
\code{get_track_lyrics} returns a tibble of lyrics and lyric metadata for a
given track. Note that currently the free API plan from Musixmatch only
returns 30% of available lyrics.
}
\examples{
\dontrun{
# Get lyrics for ...Ready for it by Taylor Swift
get_track_lyrics(134319902)
}

}
\references{
\url{https://developer.musixmatch.com/documentation/api-reference/track-lyrics-get}
}
