% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{search_artists}
\alias{search_artists}
\title{Search for a specific artist}
\usage{
search_artists(artist, page = 1, page_size = 10)
}
\arguments{
\item{artist}{string. The name of the artist to search for.}

\item{page}{numeric. The page number to return for paginated results.
Defaults to 1.}

\item{page_size}{numeric. The number of results to display per page. Defaults
to 10.}
}
\value{
A tibble containing details about matches to the artist searched for.
}
\description{
\code{search_artists} searches for a specific artist and returns data that
appears to match.
}
\examples{
\dontrun{
# Search for Taylor Swift
search_artists("Taylor Swift")
}

}
\references{
\url{https://developer.musixmatch.com/documentation/api-reference/artist-search}
}
