% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API key as option}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{character. API key obtained from \url{https://developer.musixmatch.com/signup}.}
}
\value{
Nothing.
}
\description{
\code{set_api_key} sets the musixmatch API key as an option.
}
\examples{
\dontrun{
set_api_key(YOUR_API_KEY)
}

}
